requirements.txt
langgraph
langchain
python-dotenv
langchain-openai
langchain
langchain-core
streamlit
langgraph-checkpoint-sqlite
langchain-community
uv
aiosqlite
langchain_mcp_adapters
langgraph-checkpoint-postgres
fastapi
uvicorn
psycopg[binary]